# tuma-platform/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuma API",
    "description": "Backend API for the Tuma commerce + delivery platform",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/customers": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_customers",
        "responses": {
          "200": {
            "description": "All customer accounts, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerAdminResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
    "/v1/admin/customers/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Hard-delete a customer account. The profile, refresh tokens, and order\nhistory follow via ON DELETE CASCADE — permanent, and the platform\nconfirms before calling. Scoped to customer accounts: any other identity\nis a 404.",
        "operationId": "delete_customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No customer with that id"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "operationId": "update_customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAdminResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No customer with that id"
          },
          "409": {
            "description": "Phone already taken"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/admin/merchants": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_merchants",
        "responses": {
          "200": {
            "description": "All merchant businesses, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_merchant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMerchantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Merchant business created with its owner account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedMerchantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "409": {
            "description": "Owner email already taken"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/admin/merchants/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "One business with everything the admin detail page shows: the business,\nits stores with assortment counts, and its members. Admin sees facts;\nthe merchant manages the menu — it stays out of the admin's response.",
        "operationId": "get_merchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Merchant business id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The business, its stores, and its members",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No merchant business with that id"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Hard-delete a business. Its memberships, stores (and their assortments),\nand catalog products follow via ON DELETE CASCADE. Member ACCOUNTS are\nidentities, not parts of the business — they survive the delete.",
        "operationId": "delete_merchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Merchant business id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Merchant business deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No merchant business with that id"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "operationId": "update_merchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Merchant business id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMerchantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated business",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No merchant business with that id"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/v1/admin/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Platform-wide counts for the admin dashboard.",
        "operationId": "summary",
        "responses": {
          "200": {
            "description": "Platform-wide counts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSummaryResponse": {
        "type": "object",
        "required": [
          "merchants",
          "customers",
          "stores",
          "open_stores",
          "products",
          "store_products",
          "orders_in_progress"
        ],
        "properties": {
          "customers": {
            "type": "integer",
            "format": "int64"
          },
          "merchants": {
            "type": "integer",
            "format": "int64",
            "description": "Merchant businesses."
          },
          "open_stores": {
            "type": "integer",
            "format": "int64"
          },
          "orders_in_progress": {
            "type": "integer",
            "format": "int64",
            "description": "Store orders still moving (not delivered/cancelled)."
          },
          "products": {
            "type": "integer",
            "format": "int64",
            "description": "Merchant-level catalog products."
          },
          "store_products": {
            "type": "integer",
            "format": "int64",
            "description": "Store-level sellable items."
          },
          "stores": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateMerchantInput": {
        "type": "object",
        "description": "Provision a merchant business: the business row, the owner's account\n(email + password, hashed by the AccountManager), and the owner\nmembership — one transaction. The owner signs into the merchant wing\nwith that account; the business itself never logs in.",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "business_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "CreatedMerchantResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MerchantResponse"
          },
          {
            "type": "object",
            "required": [
              "owner_email"
            ],
            "properties": {
              "owner_email": {
                "type": "string"
              }
            }
          }
        ]
      },
      "CustomerAdminResponse": {
        "type": "object",
        "description": "A customer account on the admin's list: the profile (name) joined with\nthe account facts (phone, active flag).",
        "required": [
          "user_id",
          "is_active",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "is_active": {
            "type": "boolean"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "MerchantDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MerchantResponse"
          },
          {
            "type": "object",
            "required": [
              "stores",
              "members"
            ],
            "properties": {
              "members": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              },
              "stores": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdminStoreResponse"
                }
              }
            }
          }
        ]
      },
...
      "MerchantResponse": {
        "type": "object",
        "description": "A merchant business as the admin sees it.",
        "required": [
          "id",
          "name",
          "status",
          "created_at"
        ],
        "properties": {
          "business_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "business_phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "UpdateCustomerInput": {
        "type": "object",
        "description": "Admin edits to a customer account. Provided fields overwrite, an empty\nname clears it, absent fields keep their value. Phone stays editable\nbecause a typo'd number during OTP signup is exactly the kind of thing\nan admin corrects; a taken phone is the typed 409.",
        "properties": {
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateMerchantInput": {
        "type": "object",
        "description": "Admin edits to a business. Every field is optional: provided fields\noverwrite, an empty string clears an optional contact field, absent\nfields keep their value. Suspension is the pause tool — a suspended\nbusiness is refused by the merchant wing and stops taking orders.",
        "required": [
          "status"
        ],
        "properties": {
          "business_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "business_phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}